pre-processor/requirements.txt
fastapi
uvicorn
kafka-python
opencv-python-headless
numpy
pillow 
